move workflows data to data directory
